@@ -2,32 +2,34 @@ import 'dart:convert';
 import 'dart:io';
 import 'dart:typed_data';
 
+import 'package:choice/choice.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:goport/Components/AppDialog.dart';
 import 'package:goport/Const/AppColors.dart';
+import 'package:goport/Const/AppColors.dart' as AppColors;
 import 'package:goport/Helpers/AppLocalizations.dart';
 import 'package:intl/intl.dart';
 import 'package:path_provider/path_provider.dart';
-import 'package:smart_select/smart_select.dart';
+// import 'package:smart_select/smart_select.dart';
 import 'package:unique_identifier/unique_identifier.dart';
 import 'package:fluttertoast/fluttertoast.dart';
+import 'package:choice/selection.dart';
 
 class Utils {
-  static Future<Directory> getPicturesDirectoryPath(String ext) async {
+  static Future<Directory?> getPicturesDirectoryPath(String ext) async {
     final directory = await getExternalStorageDirectory();
     return directory;
   }
 
   static Future<String> initUDID() async {
-    String identifier;
+    String? identifier;
     try {
       identifier = await UniqueIdentifier.serial;
     } on PlatformException {
       identifier = '';
     }
-
-    return identifier;
+    return identifier ?? "";
   }
 
   static showGeneralErrorToast(BuildContext context) {
@@ -47,8 +49,12 @@ class Utils {
   }
 
   static showAlertDialog(
-      {BuildContext context, String title = "", String message, Function onOk}) {
-    AppDialog alert = AppDialog(title: title, message: message, onOK: onOk,);
+      {required BuildContext context,
+      String title = "",
+      String? message,
+      Function? onOk}) {
+    AppDialog alert =
+        AppDialog(title: title, message: message, onOK: () => onOk!());
 
     showDialog(
       context: context,
@@ -58,14 +64,15 @@ class Utils {
     );
   }
 
-  static showConfirmDialog(
-      {BuildContext context,
-      String title,
-      String message,
-      String okButton,
-      String cancelButton,
-        VoidCallback onCancel,
-      VoidCallback onOk}) {
+  static showConfirmDialog({
+    BuildContext? context,
+    String? title,
+    required String message,
+    required String okButton,
+    String? cancelButton,
+    VoidCallback? onCancel,
+    required VoidCallback onOk,
+  }) {
     AppDialog alert = AppDialog(
       title: title,
       message: message,
@@ -76,37 +83,98 @@ class Utils {
     );
 
     showDialog(
-        context: context,
-        builder: (BuildContext context) {
-          return alert;
-        });
+      context: context!,
+      builder: (BuildContext context) {
+        return alert;
+      },
+    );
   }
 
-  static showSingleChoiceDialog(
-      {BuildContext context,
-      String title,
-      List<String> options,
-      String selected,
-      Function onSelect}) {
-    List<S2Choice<String>> choiceItems =
-        options.map((e) => S2Choice<String>(value: e, title: e)).toList();
-    return SmartSelect<String>.single(
-        modalType: S2ModalType.popupDialog,
-        tileBuilder: (context, state) {
-          return ListTile(
-            title: Text(selected ?? state.title, textAlign: TextAlign.center, style: TextStyle(
-              color: colorLogo2, fontWeight: FontWeight.bold
-            ),),
-            onTap: state.showModal,
-          );
-        },
-        title: title,
-        value: selected,
-        choiceItems: choiceItems,
-        onChange: (state) {
-          int index = choiceItems.indexOf(state.valueObject);
-          onSelect(index, state.value);
-        });
+  // static showSingleChoiceDialog({
+  //   required BuildContext context,
+  //   required String title,
+  //   required List<String> options,
+  //   required String selected,
+  //   required Function onSelect,
+  // }) {
+  //   List<S2Choice<String>> choiceItems =
+  //       options.map((e) => S2Choice<String>(value: e, title: e)).toList();
+  //   return SmartSelect<String>.single(
+  //     modalType: S2ModalType.popupDialog,
+  //     tileBuilder: (context, state) {
+  //       return ListTile(
+  //         title: Text(
+  //           selected,
+  //           textAlign: TextAlign.center,
+  //           style: const TextStyle(
+  //             color: AppColors.colorLogo2,
+  //             fontWeight: FontWeight.bold,
+  //           ),
+  //         ),
+  //         onTap: state.showModal,
+  //       );
+  //     },
+  //     title: title,
+  //     selectedValue: selected,
+  //     choiceItems: choiceItems,
+  //     onChange: (state) {
+  //       // Compute index by matching the selected value with the choice items
+  //       int index = choiceItems.indexWhere((item) => item.value == state.value);
+  //       onSelect(index, state.value);
+  //     },
+  //   );
+  // }
+  static showSingleChoiceDialog({
+    required BuildContext context,
+    required String title,
+    required List<String> options,
+    required String selected,
+    required Function onSelect,
+  }) {
+    return showDialog(
+      context: context,
+      builder: (BuildContext context) {
+        return AlertDialog(
+          title: Text(
+            title,
+            textAlign: TextAlign.center,
+          ),
+          content: Container(
+            width: double.maxFinite,
+            child: PromptedChoice<String>.single(
+              value: selected,
+              onChanged: (value) {
+                // Compute index by matching the selected value with the options
+                int index = options.indexOf(value!);
+                onSelect(index, value);
+                Navigator.of(context).pop();
+              },
+              itemBuilder: (ChoiceController<String> state, int index) {
+                String option = options[index];
+                return RadioListTile<String>(
+                  title: Text(option),
+                  value: option,
+                  groupValue: state.single,
+                  onChanged: (value) {
+                    state.select(options[index]);
+                  },
+                  selected: option == state.value,
+                );
+              },
+              itemCount: options.length,
+            ),
+          ),
+          actions: [
+            TextButton(
+              onPressed: () {
+                Navigator.of(context).pop();
+              },
+              child: Text('Cancel'),
+            ),
+          ],
+        );
+      },
+    );
   }
 
   static String convertToBase64String(String input) {
@@ -115,7 +183,7 @@ class Utils {
   }
 
   static String convertImageToBase64(String imagePath) {
-    File imageFile = new File(imagePath.toString());
+    File imageFile = File(imagePath);
     List<int> imageBytes = imageFile.readAsBytesSync();
     return base64Encode(imageBytes);
   }
