@@ -4,16 +4,16 @@ part 'GoPortAppStatus.g.dart';
 @JsonSerializable()
 class GoPortAppStatus {
   final int id;
-  final String driverTZ;
-  final String guidID;
-  final String truckNum;
-  final String trailerNum;
-  final bool appInForeground;
-  final bool driverAccepted;
-  final DateTime timeStamp;
+  final String? driverTZ;
+  final String? guidID;
+  final String? truckNum;
+  final String? trailerNum;
+  final bool? appInForeground;
+  final bool? driverAccepted;
+  final DateTime? timeStamp;
 
   GoPortAppStatus(
-      {this.id,
+      {required this.id,
       this.driverTZ,
       this.guidID,
       this.truckNum,
@@ -22,7 +22,7 @@ class GoPortAppStatus {
       this.driverAccepted,
       this.timeStamp});
 
-  factory GoPortAppStatus.fromJson(Map<String, dynamic> json) => _$GoPortAppStatusFromJson(json);
+  factory GoPortAppStatus.fromJson(Map<String, dynamic> json) =>
+      _$GoPortAppStatusFromJson(json);
   Map<String, dynamic> toJson() => _$GoPortAppStatusToJson(this);
-
 }
