@@ -4,35 +4,35 @@ part 'PortContainer.g.dart';
 @JsonSerializable()
 class PortContainer {
   int id;
-  String actualCntrNo;
-  String containerType;
-  double weight;
-  String fullEmptyContainer;
-  String bay;
-  String ca;
-  String size;
-  String block;
-  String declaredSealNo;
-  String informationStatus;
-  String pickUpPlanTime;
-  String pickUpPlanTime2;
-  String remarks;
-  String positionContainer;
-  String shipperCompanyName;
-  String shippingAgent;
-  String sealNo;
-  int draftID;
-  String plombaNumber;
-  String imagePath;
-  String imageName;
-  String image;
-  String lat;
-  String lng;
-  String deliveryDocType;
-  String cargoType;
+  String? actualCntrNo;
+  String? containerType;
+  double? weight;
+  String? fullEmptyContainer;
+  String? bay;
+  String? ca;
+  String? size;
+  String? block;
+  String? declaredSealNo;
+  String? informationStatus;
+  String? pickUpPlanTime;
+  String? pickUpPlanTime2;
+  String? remarks;
+  String? positionContainer;
+  String? shipperCompanyName;
+  String? shippingAgent;
+  String? sealNo;
+  int? draftID;
+  String? plombaNumber;
+  String? imagePath;
+  String? imageName;
+  String? image;
+  String? lat;
+  String? lng;
+  String? deliveryDocType;
+  String? cargoType;
 
   PortContainer(
-  {this.id,
+      {required this.id,
       this.actualCntrNo,
       this.containerType,
       this.weight,
@@ -60,6 +60,7 @@ class PortContainer {
       this.deliveryDocType,
       this.cargoType});
 
-  factory PortContainer.fromJson(Map<String, dynamic> json) => _$PortContainerFromJson(json);
+  factory PortContainer.fromJson(Map<String, dynamic> json) =>
+      _$PortContainerFromJson(json);
   Map<String, dynamic> toJson() => _$PortContainerToJson(this);
 }
