@@ -1,9 +1,12 @@
 import 'package:flutter/material.dart';
 
 class RootDrawer {
-  static DrawerControllerState of(BuildContext context) {
-    final DrawerControllerState drawerControllerState =
-    context.findRootAncestorStateOfType();
-    return drawerControllerState;
+  static DrawerControllerState? of(BuildContext context) {
+    // Method returns State<StatefulWidget>? which is nullable
+    final State<StatefulWidget>? state =
+        context.findRootAncestorStateOfType<DrawerControllerState>();
+
+    // Cast to the correct type if not null
+    return state as DrawerControllerState?;
   }
 }
