@@ -6,20 +6,15 @@ part of 'AvailableJobs.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-AvailableJobs _$AvailableJobsFromJson(Map<String, dynamic> json) {
-  return AvailableJobs(
-    containerJobsIn: (json['containerJobsIn'] as List)
-        ?.map((e) => e == null
-            ? null
-            : PortContainer.fromJson(e as Map<String, dynamic>))
-        ?.toList(),
-    containerJobsOut: (json['containerJobsOut'] as List)
-        ?.map((e) => e == null
-            ? null
-            : PortContainer.fromJson(e as Map<String, dynamic>))
-        ?.toList(),
-  );
-}
+AvailableJobs _$AvailableJobsFromJson(Map<String, dynamic> json) =>
+    AvailableJobs(
+      containerJobsIn: (json['containerJobsIn'] as List<dynamic>)
+          .map((e) => PortContainer.fromJson(e as Map<String, dynamic>))
+          .toList(),
+      containerJobsOut: (json['containerJobsOut'] as List<dynamic>)
+          .map((e) => PortContainer.fromJson(e as Map<String, dynamic>))
+          .toList(),
+    );
 
 Map<String, dynamic> _$AvailableJobsToJson(AvailableJobs instance) =>
     <String, dynamic>{
