@@ -6,22 +6,20 @@ part of 'DraftJobCard.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-DraftJobCard _$DraftJobCardFromJson(Map<String, dynamic> json) {
-  return DraftJobCard(
-    id: json['id'] as int,
-    driverTZ: json['driverTZ'] as String,
-    containerNo: json['containerNo'] as String,
-    createDate: json['createDate'] == null
-        ? null
-        : DateTime.parse(json['createDate'] as String),
-    containerJobTypeID: json['containerJobTypeID'] as int,
-    plombaNumber: json['plombaNumber'] as String,
-    imagePath: json['imagePath'] as String,
-    imageName: json['imageName'] as String,
-    image: json['image'] as String,
-    notTakePhotoReasonID: json['notTakePhotoReasonID'] as int,
-  );
-}
+DraftJobCard _$DraftJobCardFromJson(Map<String, dynamic> json) => DraftJobCard(
+      id: (json['id'] as num).toInt(),
+      driverTZ: json['driverTZ'] as String?,
+      containerNo: json['containerNo'] as String?,
+      createDate: json['createDate'] == null
+          ? null
+          : DateTime.parse(json['createDate'] as String),
+      containerJobTypeID: (json['containerJobTypeID'] as num?)?.toInt(),
+      plombaNumber: json['plombaNumber'] as String?,
+      imagePath: json['imagePath'] as String?,
+      imageName: json['imageName'] as String?,
+      image: json['image'] as String?,
+      notTakePhotoReasonID: (json['notTakePhotoReasonID'] as num?)?.toInt(),
+    );
 
 Map<String, dynamic> _$DraftJobCardToJson(DraftJobCard instance) =>
     <String, dynamic>{
