@@ -6,37 +6,36 @@ part of 'PortContainer.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-PortContainer _$PortContainerFromJson(Map<String, dynamic> json) {
-  return PortContainer(
-    id: json['id'] as int,
-    actualCntrNo: json['actualCntrNo'] as String,
-    containerType: json['containerType'] as String,
-    weight: (json['weight'] as num)?.toDouble(),
-    fullEmptyContainer: json['fullEmptyContainer'] as String,
-    bay: json['bay'] as String,
-    ca: json['ca'] as String,
-    size: json['size'] as String,
-    block: json['block'] as String,
-    declaredSealNo: json['declaredSealNo'] as String,
-    informationStatus: json['informationStatus'] as String,
-    pickUpPlanTime: json['pickUpPlanTime'] as String,
-    pickUpPlanTime2: json['pickUpPlanTime2'] as String,
-    remarks: json['remarks'] as String,
-    positionContainer: json['positionContainer'] as String,
-    shipperCompanyName: json['shipperCompanyName'] as String,
-    shippingAgent: json['shippingAgent'] as String,
-    sealNo: json['sealNo'] as String,
-    draftID: json['draftID'] as int,
-    plombaNumber: json['plombaNumber'] as String,
-    imagePath: json['imagePath'] as String,
-    imageName: json['imageName'] as String,
-    image: json['image'] as String,
-    lat: json['lat'] as String,
-    lng: json['lng'] as String,
-    deliveryDocType: json['deliveryDocType'] as String,
-    cargoType: json['cargoType'] as String,
-  );
-}
+PortContainer _$PortContainerFromJson(Map<String, dynamic> json) =>
+    PortContainer(
+      id: (json['id'] as num).toInt(),
+      actualCntrNo: json['actualCntrNo'] as String?,
+      containerType: json['containerType'] as String?,
+      weight: (json['weight'] as num?)?.toDouble(),
+      fullEmptyContainer: json['fullEmptyContainer'] as String?,
+      bay: json['bay'] as String?,
+      ca: json['ca'] as String?,
+      size: json['size'] as String?,
+      block: json['block'] as String?,
+      declaredSealNo: json['declaredSealNo'] as String?,
+      informationStatus: json['informationStatus'] as String?,
+      pickUpPlanTime: json['pickUpPlanTime'] as String?,
+      pickUpPlanTime2: json['pickUpPlanTime2'] as String?,
+      remarks: json['remarks'] as String?,
+      positionContainer: json['positionContainer'] as String?,
+      shipperCompanyName: json['shipperCompanyName'] as String?,
+      shippingAgent: json['shippingAgent'] as String?,
+      sealNo: json['sealNo'] as String?,
+      draftID: (json['draftID'] as num?)?.toInt(),
+      plombaNumber: json['plombaNumber'] as String?,
+      imagePath: json['imagePath'] as String?,
+      imageName: json['imageName'] as String?,
+      image: json['image'] as String?,
+      lat: json['lat'] as String?,
+      lng: json['lng'] as String?,
+      deliveryDocType: json['deliveryDocType'] as String?,
+      cargoType: json['cargoType'] as String?,
+    );
 
 Map<String, dynamic> _$PortContainerToJson(PortContainer instance) =>
     <String, dynamic>{
