@@ -6,14 +6,11 @@ part of 'ShipResponse.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-ShipResponse _$ShipResponseFromJson(Map<String, dynamic> json) {
-  return ShipResponse(
-    (json['shipsList'] as List)
-        ?.map(
-            (e) => e == null ? null : Ship.fromJson(e as Map<String, dynamic>))
-        ?.toList(),
-  );
-}
+ShipResponse _$ShipResponseFromJson(Map<String, dynamic> json) => ShipResponse(
+      (json['shipsList'] as List<dynamic>)
+          .map((e) => Ship.fromJson(e as Map<String, dynamic>))
+          .toList(),
+    );
 
 Map<String, dynamic> _$ShipResponseToJson(ShipResponse instance) =>
     <String, dynamic>{
