@@ -6,14 +6,12 @@ part of 'WarehouseResponse.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-WarehouseResponse _$WarehouseResponseFromJson(Map<String, dynamic> json) {
-  return WarehouseResponse(
-    (json['machsanList'] as List)
-        ?.map((e) =>
-            e == null ? null : Warehouse.fromJson(e as Map<String, dynamic>))
-        ?.toList(),
-  );
-}
+WarehouseResponse _$WarehouseResponseFromJson(Map<String, dynamic> json) =>
+    WarehouseResponse(
+      (json['machsanList'] as List<dynamic>)
+          .map((e) => Warehouse.fromJson(e as Map<String, dynamic>))
+          .toList(),
+    );
 
 Map<String, dynamic> _$WarehouseResponseToJson(WarehouseResponse instance) =>
     <String, dynamic>{
