@@ -7,15 +7,13 @@ part of 'SaveContainersToDraft.dart';
 // **************************************************************************
 
 SaveContainersToDraft _$SaveContainersToDraftFromJson(
-    Map<String, dynamic> json) {
-  return SaveContainersToDraft(
-    driverTZ: json['driverTZ'] as String,
-    containersToAdd: (json['containersToAdd'] as List)
-        ?.map((e) =>
-            e == null ? null : DraftJobCard.fromJson(e as Map<String, dynamic>))
-        ?.toList(),
-  );
-}
+        Map<String, dynamic> json) =>
+    SaveContainersToDraft(
+      driverTZ: json['driverTZ'] as String,
+      containersToAdd: (json['containersToAdd'] as List<dynamic>)
+          .map((e) => DraftJobCard.fromJson(e as Map<String, dynamic>))
+          .toList(),
+    );
 
 Map<String, dynamic> _$SaveContainersToDraftToJson(
         SaveContainersToDraft instance) =>
