@@ -6,20 +6,18 @@ part of 'Ship.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-Ship _$ShipFromJson(Map<String, dynamic> json) {
-  return Ship(
-    json['ship_NO'] as String,
-    json['shipnamE_ENG'] as String,
-    json['stat'] as String,
-    json['sochen'] as String,
-    json['razif'] as String,
-    json['lasT_UPDATE'] as String,
-    json['x_GPS'] as String,
-    json['y_GPS'] as String,
-    json['zakef'] as String,
-    json['selected'] as bool,
-  );
-}
+Ship _$ShipFromJson(Map<String, dynamic> json) => Ship(
+      json['ship_NO'] as String,
+      json['shipnamE_ENG'] as String,
+      json['stat'] as String,
+      json['sochen'] as String,
+      json['razif'] as String,
+      json['lasT_UPDATE'] as String,
+      json['x_GPS'] as String,
+      json['y_GPS'] as String,
+      json['zakef'] as String,
+      json['selected'] as bool,
+    );
 
 Map<String, dynamic> _$ShipToJson(Ship instance) => <String, dynamic>{
       'ship_NO': instance.ship_NO,
