@@ -6,15 +6,13 @@ part of 'Technician.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-Technician _$TechnicianFromJson(Map<String, dynamic> json) {
-  return Technician(
-    json['id'] as int,
-    json['containerTypeID'] as String,
-    json['technicianName'] as String,
-    json['phoneNumber'] as String,
-    json['address'] as String,
-  );
-}
+Technician _$TechnicianFromJson(Map<String, dynamic> json) => Technician(
+      (json['id'] as num).toInt(),
+      json['containerTypeID'] as String,
+      json['technicianName'] as String,
+      json['phoneNumber'] as String,
+      json['address'] as String,
+    );
 
 Map<String, dynamic> _$TechnicianToJson(Technician instance) =>
     <String, dynamic>{
