@@ -6,14 +6,12 @@ part of 'Warehouse.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-Warehouse _$WarehouseFromJson(Map<String, dynamic> json) {
-  return Warehouse(
-    json['shemMachsan'] as String,
-    json['x_GPS'] as String,
-    json['y_GPS'] as String,
-    json['selected'] as bool,
-  );
-}
+Warehouse _$WarehouseFromJson(Map<String, dynamic> json) => Warehouse(
+      json['shemMachsan'] as String,
+      json['x_GPS'] as String,
+      json['y_GPS'] as String,
+      json['selected'] as bool,
+    );
 
 Map<String, dynamic> _$WarehouseToJson(Warehouse instance) => <String, dynamic>{
       'shemMachsan': instance.shemMachsan,
